requirements.txt
requests
statsd
responder
PyJWT
cryptography
